requirements.txt
# Windows-compatible requirements with pre-compiled wheels
# Core LLM and RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0

# Vector Database (use older version for Windows compatibility)
chromadb>=0.4.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
python-pptx>=0.6.20
docx2txt
tiktoken

# Embeddings (use compatible version)
sentence-transformers>=2.2.0

# Web Interface - Streamlit Only
streamlit>=1.30.0

# Web Search
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=4.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0
tqdm>=4.65.0

# Essential dependencies
numpy>=1.24.0
scipy>=1.10.0